add README for cli testing; move Mermaid functions to separate file
